+++ TentakelsAttacking2/Game/public/GameManager.h
@@ -9,7 +9,7 @@
 #include "EventListener.hpp"
 #include "UIEvents.hpp"
 #include "GenerelEvents.hpp"
-#include "GameEventTypes.hpp"
+#include "HGameEventTypes.hpp"
 #include <vector>
 #include <random>
 #include <memory>
